+++ src/com/kabe/donhackday2014/TakePhotoActivity.java
@@ -87,8 +87,8 @@ public class TakePhotoActivity extends Activity {
 		Matrix matrix = new Matrix();
 		DisplayMetrics metrics = new DisplayMetrics();
 		activity.getWindowManager().getDefaultDisplay().getMetrics(metrics);
-		float screenWidth = (float) metrics.widthPixels;
-		float screenHeight = (float) metrics.heightPixels;
+		float screenWidth = metrics.widthPixels;
+		float screenHeight = metrics.heightPixels;
 		Log.d(TAG, "screenWidth = " + String.valueOf(screenWidth)
 				+ " px, screenHeight = " + String.valueOf(screenHeight) + " px");
